// ui/src/components/__generated__/PostArticle.ts
/* tslint:disable */
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL mutation operation: PostArticle
// ====================================================

export interface PostArticle_postArticle {
  id: string;
  title: string;
  content: string;
}

export interface PostArticle {
  postArticle: PostArticle_postArticle;
}

export interface PostArticleVariables {
  diaryId: string;
  title: string;
  content: string;
}

/* tslint:disable */
// This file was automatically generated and should not be edited.

//==============================================================
// START Enums and Input Objects
//==============================================================

//==============================================================
// END Enums and Input Objects
//==============================================================